+++ OUDS/Core/Components/Sources/Indicators/Badge/Internal/BadgeCount.swift
@@ -37,14 +37,3 @@ struct BadgeCount: View {
         }
     }
 }
-
-extension OUDSBadge.IllustrationSize {
-    private var countSize: OUDSBadgeCount.Size {
-        switch self {
-        case .medium:
-            .medium
-        case .large:
-            .large
-        }
-    }
-}