+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- **BREAKING**: Deprecated `OUDSBadge` API
 - **BREAKING**: Deprecated type `OUDSIcon`
 - **BREAKING**: Deprecated initializers for `button`, `checkbox`, `chips`, `radio`, `switch`, `checkbox`, `text input`, `badge`, `link`, `tag` components
 - **BREAKING**: Component tokens of `control item` component (Orange-OpenSource/ouds-ios#1579)